% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bmi}
\alias{bmi}
\title{BMI formula}
\usage{
bmi(mass, height)
}
\arguments{
\item{mass}{Numeric. Mass in Kg.}

\item{height}{Numeric. Height in m.}
}
\value{
numeric.
}
\description{
BMI formula
}
\seealso{
Other utils: 
\code{\link{anemia}()},
\code{\link{egfr}()},
\code{\link{fib4}()},
\code{\link{low_bmi}()}
}
\concept{utils}
