% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\alias{conv_g.ml_to_g.dl}
\alias{conv_g.dl_to_g.ml}
\alias{conv_mul_to_mm3}
\alias{conv_mm3_to_mul}
\alias{conv_K.ml_to_K.mm3}
\alias{conv_K.mm3_to_K.ml}
\alias{conv_cm_to_m}
\alias{conv_m_to_cm}
\alias{conv_mumol.l_to_mg.dl}
\alias{conv_mg.dl_to_mumol.l}
\title{Conversion of units.}
\usage{
conv_g.ml_to_g.dl(x)

conv_g.dl_to_g.ml(x)

conv_mul_to_mm3(x)

conv_mm3_to_mul(x)

conv_K.ml_to_K.mm3(x)

conv_K.mm3_to_K.ml(x)

conv_cm_to_m(x)

conv_m_to_cm(x)

conv_mumol.l_to_mg.dl(x)

conv_mg.dl_to_mumol.l(x)
}
\arguments{
\item{x}{Numeric.}
}
\value{
A numerical vector.
}
\description{
Conversion of units.
}
\details{
\code{conv_g.ml_to_g.dl()}: g/ml to g/dl. \eqn{1/100}.

\code{conv_g.dl_to_g.ml()}: g/dl to g/ml. x\eqn{100}.

\code{conv_mul_to_mm3()}: \eqn{\mu}l to mm\eqn{^3}. x\eqn{1}.

\code{conv_mm3_to_ml()}: mm\eqn{^3} to \eqn{\mu}l. x\eqn{1}.

\code{conv_K.ml_to_K.mm3()}. count per ml to count per mm\eqn{^3}. \eqn{1/1000}.

\code{conv_K.mm3_to_K.ml()}. count per mm\eqn{^3} to count per mL. x\eqn{1000}.

\code{conv_cm_to_m()}. cm to m. \eqn{1/100}.

\code{conv_cm_to_m()}. m to cm. x\eqn{100}.

\code{conv_mumol.l_to_mg.dl()}. \eqn{\mu} mod per l to mg per dl. x \eqn{0.0113096584483149}

\code{conv_mg.dl_to_mumol.l()}. mg per dl to \eqn{\mu} mod per l. \eqn{1/0.0113096584483149}
}
\concept{convert}
