% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataids}
\alias{dataids}
\title{Test data the DAT'aids score}
\format{
A data frame with 1152 rows and 9 variables:
\describe{
  \item{age}{Numeric. Age.}
  \item{cd4_count}{Numeric. CD4 count.}
  \item{nonhiv_cancer}{Logical. History of non-hiv cancer.}
  \item{cardio_disease}{Logical. History of cardiac disease.}
  \item{egfr}{Numeric. eGFR.}
  \item{cirrhosis}{Logical. History of cirrhosis.}
  \item{low_bmi}{Logical. Low BMI.}
  \item{anemia}{Logical. Anemia.}
  \item{SCORE_DATAIDS}{Numeric. Score Dat'Aids}
}
}
\source{
\url{https://doi.org/10.1371/journal.pone.0195725/}
}
\usage{
dataids
}
\description{
A dataset for testing the DAT'aids score
}
\keyword{datasets}
