% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataids.R
\name{dataids_score}
\alias{dataids_score}
\alias{surv_dataids}
\alias{eta_dataids}
\alias{eta_riskgroupdataids}
\alias{score_dataids}
\alias{riskgroup_dataids}
\alias{points_dataids}
\title{Dat'AIDS score}
\usage{
surv_dataids(
  time,
  score = NULL,
  basesurv_fun = basesurv_hentzien2018,
  eta_fun = eta_dataids,
  age = NULL,
  cd4_count = NULL,
  nonhiv_cancer = NULL,
  cardio_disease = NULL,
  egfr = NULL,
  cirrhosis = NULL,
  low_bmi = NULL,
  anemia = NULL
)

eta_dataids(
  score = NULL,
  beta_score = 0.1008354015,
  age = NULL,
  cd4_count = NULL,
  nonhiv_cancer = NULL,
  cardio_disease = NULL,
  egfr = NULL,
  cirrhosis = NULL,
  low_bmi = NULL,
  anemia = NULL
)

eta_riskgroupdataids(
  score = NULL,
  beta_score = c(`0-3` = 0, `4-13` = 0.81868, `14-19` = 1.76034, `>=20` = 2.61992),
  age = NULL,
  cd4_count = NULL,
  nonhiv_cancer = NULL,
  cardio_disease = NULL,
  egfr = NULL,
  cirrhosis = NULL,
  low_bmi = NULL,
  anemia = NULL
)

score_dataids(
  age,
  cd4_count,
  nonhiv_cancer,
  cardio_disease,
  egfr,
  cirrhosis,
  low_bmi,
  anemia
)

riskgroup_dataids(score)

points_dataids()
}
\arguments{
\item{time}{Numeric. Time of prediction in year.}

\item{score}{Numeric. Dat'AIDS score. If \code{NULL}, the score is computed using the predictors.}

\item{basesurv_fun}{Function of \code{time}. Default is \code{basesurv_hentzien2018}.}

\item{eta_fun}{Function of the score which returns the linear predictor. Default is \code{eta_dataids} and \code{eta_riskgroupdataids} is available.}

\item{age}{Numeric. Age in year.}

\item{cd4_count}{Numeric. Count of CD4 cell/mm\eqn{^3}.}

\item{nonhiv_cancer}{Logical. \code{TRUE} for diagnosis of non-HIV cancer.}

\item{cardio_disease}{Logical. \code{TRUE} for diagnosis of cardiovascular disease.}

\item{egfr}{Numeric. Estimated glomerular filtration rate. (ml/min/1.73m\eqn{^2}).}

\item{cirrhosis}{Logical. \code{TRUE} for cirrhosis diagnostic.}

\item{low_bmi}{Logical. \code{TRUE} if BMI < 18.5.}

\item{anemia}{Logical. \code{TRUE} for anemia defined as hemoglobin <12g/dL for female and <13g/dL for male.}

\item{beta_score}{Numberic. Default is \code{0.1008354015} for continuous score and \code{c("0-3"=0, "4-13"=0.81868, "14-19"=1.76034, ">=20"=2.61992)} for risk group. Parameter associated to the score.}
}
\value{
Numeric. Predicted survival probability.
}
\description{
Compute the Dat'AIDS score and survival prediction
}
\details{
The prediction of survival at time t is:
\deqn{\hat{S}(t) = S_0(t)^{\exp{b*\text{score}}}},
where \eqn{S_0(t)} is \code{basesurv_fun} and \eqn{b} is \code{beta_score}.

\code{score_dataids()} returns a numerical vector, \code{points_dataids()} returns a list of the parameters used to compute the VACS Index 1 and \code{surv_dataids()} returns the survival probabilities a the baseline survival function.
}
\examples{
data(dataids)
dataids$dataids <- score_dataids(
    age = dataids$age, cd4_count = dataids$cd4_count,
    nonhiv_cancer = dataids$nonhiv_cancer, cardio_disease = dataids$cardio_disease,
    egfr = dataids$egfr, cirrhosis = dataids$cirrhosis, low_bmi = dataids$low_bmi,
    anemia = dataids$anemia)
}
\references{
Hentzien M, Delpierre C, Pugliese P, Allavena C, Jacomet C, Valantin M-A, et al. (2018) Derivation and internal validation of a mortality risk index for aged people living with HIV: The Dat'AIDS score. PLoS ONE 13(4): e0195725. https://doi.org/10.1371/journal.pone.0195725
}
\seealso{
Other dataids: 
\code{\link{baseline}}

Other dataids: 
\code{\link{baseline}}

Other dataids: 
\code{\link{baseline}}
}
\concept{dataids}
