% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{discretize}
\alias{discretize}
\alias{discrete_age_dataids}
\alias{discrete_age_vacs1}
\alias{discrete_cd4_dataids}
\alias{discrete_cd4_vacs1}
\alias{discrete_egfr_dataids}
\alias{discrete_egfr_vacs1}
\alias{discrete_fib4_vacs1}
\alias{discrete_hgb_vacs1}
\alias{discrete_hiv1rna_vacs1}
\title{Discretization of numeric variable used in the DAT'aids and VACS Index 1}
\usage{
discrete_age_dataids(x)

discrete_age_vacs1(x)

discrete_cd4_dataids(x)

discrete_cd4_vacs1(x)

discrete_egfr_dataids(x)

discrete_egfr_vacs1(x)

discrete_fib4_vacs1(x)

discrete_hgb_vacs1(x)

discrete_hiv1rna_vacs1(x)
}
\arguments{
\item{x}{Numeric variable to discretize.}
}
\value{
Factor.
}
\description{
Discretization of numeric variable used in the DAT'aids and VACS Index 1
}
\examples{
data(dataids)
dataids_d<-
  data.frame(
    age = dataids$age,
    age_d = discrete_age_dataids(dataids$age),
    cd4 = dataids$cd4_count,
    cd4_d = discrete_cd4_dataids(dataids$cd4_count),
    egfr = dataids$egfr,
    egfr_d = discrete_egfr_dataids(dataids$egfr))
head(dataids_d)

data(vacs2)
vacs2_d <-
  data.frame(
    age = vacs2$AGE,
    age_d = discrete_age_vacs1(vacs2$AGE),
    cd4 = vacs2$CD4,
    cd4_d = discrete_cd4_vacs1(vacs2$CD4),
    vl = 10^vacs2$VL_LOG,
    vl_d = discrete_hiv1rna_vacs1(10^vacs2$VL_LOG),
    egfr = vacs2$EGFR,
    egfr_d = discrete_egfr_vacs1(vacs2$EGFR),
    fib4 = vacs2$EGFR,
    fib4_d = discrete_fib4_vacs1(vacs2$FIB4),
    hgb = vacs2$HGB,
    hgb_d = discrete_fib4_vacs1(vacs2$HGB))
head(vacs2_d)
}
