% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{egfr}
\alias{egfr}
\alias{egfr_norace}
\title{Estimated glomerular filtration rate formula}
\usage{
egfr(scr, age, sex, race)

egfr_norace(scr, age, sex)
}
\arguments{
\item{scr}{Numeric. Serum creatinine \eqn{\mu} mol/l}

\item{age}{Numeric. Year.}

\item{sex}{Character.}

\item{race}{Character.}
}
\value{
A numeric
}
\description{
Estimated glomerular filtration rate formula
}
\details{
The eGFR formula are is derived from

Levey, A. S., Stevens, L. A., Schmid, C. H., Zhang, Y., Castro III, A. F., Feldman, H. I., ... & CKD-EPI (Chronic Kidney Disease Epidemiology Collaboration)*. (2009). A new equation to estimate glomerular filtration rate. Annals of internal medicine, 150(9), 604-612.
}
