% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{low_bmi}
\alias{low_bmi}
\title{Low bmi}
\usage{
low_bmi(bmi)
}
\arguments{
\item{bmi}{Numeric. Body Mass Index.}
}
\value{
Logical. TRUE if bmi < 18.5, FALSE otherwise.
}
\description{
Low bmi
}
\seealso{
Other utils: 
\code{\link{anemia}()},
\code{\link{bmi}()},
\code{\link{egfr}()},
\code{\link{fib4}()},
\code{\link{vl_log}()}
}
\concept{utils}
