% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vacs1}
\alias{vacs1}
\title{Testing data for the VACS Index 1 and VACS Index 2}
\format{
A data frame with 84 rows and 13 variables:
\describe{
  \item{age}{Integer. Years.}
  \item{cd4}{Integer. CD4 count. K/mm^3.}
  \item{log_hiv1_rna}{Numeric. HIV1 RNA. log K/ml.}
  \item{hgb}{Integer. Hemoglobin. g/dL.}
  \item{fib4}{Numeric. FIB-4.}
  \item{egfr}{Integer. eGFR.}
  \item{hepc}{Logical. Hepatitis C co-infection.}
  \item{alb}{Integer. Albumin. g/dL.}
  \item{wbc}{Numeric. White blood cell count. K/mL.}
  \item{bmi}{Numeric. Body mass Index. kg/m^2}
  \item{hiv1_rna}{Numeric. HIV1 RNA. K/ml.}
  \item{vacs1}{Integer. Score Vacs Index 1.}
  \item{vacs2}{Numeric. Score Vacs Index 2.}
}
}
\usage{
vacs1
}
\description{
A simulated dataset
}
\keyword{datasets}
