% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vacs2}
\alias{vacs2}
\title{Testing data for the VACS Index 2}
\format{
A data frame with 94 rows and 13 variables:
\describe{
  \item{PARAM}{Numeric.}
  \item{SET}{Numeric.}
  \item{AGE}{Numeric. Age.}
  \item{CD4}{Numeric. CD4 count.}
  \item{VL_LOG}{Numeric. Log of viral load.}
  \item{HGB}{Numeric. Hemoglobin.}
  \item{FIB4}{Numeric. FIB-4.}
  \item{EGFR}{Numeric. eGFR.}
  \item{HEPC}{Numeric. Hepatitis C.}
  \item{ALB}{Numeric. Albumine.}
  \item{WBC}{Numeric. White Blood Cell.}
  \item{BMI}{Numeric. BMI.}
  \item{SCORE_V2}{Vacs Index 2.}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/34609485/}
}
\usage{
vacs2
}
\description{
A dataset containing the testing data in the supplementary material of doi: 10.1093/cid/ciab883
}
\keyword{datasets}
