% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacsindex1.R
\name{vacsindex1}
\alias{vacsindex1}
\alias{surv_vacsindex1}
\alias{score_vacsindex1}
\alias{points_vacsindex1}
\title{Compute the VACS Index 1 survival prediction}
\usage{
surv_vacsindex1(
  time,
  baseline_fun,
  score = NULL,
  age = NULL,
  cd4_count = NULL,
  hiv1_rna = NULL,
  hgb = NULL,
  fib4 = NULL,
  egfr = NULL,
  hepc = NULL
)

score_vacsindex1(age, cd4_count, hiv1_rna, hgb, fib4, egfr, hepc)

points_vacsindex1()
}
\arguments{
\item{time}{Numeric. Time of prediction in year.}

\item{baseline_fun}{Function of \code{time}.}

\item{score}{Numeric. VACS index 1 score. If NULL, it will be computed based on the covariates.}

\item{age}{Numeric. Years}

\item{cd4_count}{Numeric. CD4 cell count. cell/mm^3}

\item{hiv1_rna}{Numeric. Viral load. copies/mL}

\item{hgb}{Numeric. Hemoglobin. g/dL.}

\item{fib4}{Numeric. Fibrosis-4 Index for Liver Fibrosis.}

\item{egfr}{Numeric. Estimated glomerular filtration rate. (ml/mn/1.73m2).}

\item{hepc}{Logical. TRUE for co-infection with hepatitis C virus.}
}
\value{
A numerical vector.
}
\description{
Compute the VACS Index 1 survival prediction
}
\references{
Tate, J. P., Justice, A. C., Hughes, M. D., Bonnet, F., Reiss, P., Mocroft, A., ... & Sterne, J. A. (2013). An internationally generalizable risk index for mortality after one year of antiretroviral therapy. AIDS (London, England), 27(4), 563.

Bebu, I., Tate, J., Rimland, D., Mesner, O., Macalino, G. E., Ganesan, A., ... & Infectious Disease Clinical Research Program HIV Working Group. (2014). The VACS index predicts mortality in a young, healthy HIV population starting highly active antiretroviral therapy. Journal of acquired immune deficiency syndromes (1999), 65(2), 226.

Justice, A. C., Modur, S., Tate, J. P., Althoff, K. N., Jacobson, L. P., Gebo, K., ... & Gange, S. J. (2013). Predictive accuracy of the Veterans Aging Cohort Study (VACS) index for mortality with HIV infection: a north American cross cohort analysis. Journal of acquired immune deficiency syndromes (1999), 62(2), 149.
}
