% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vl_log}
\alias{vl_log}
\title{Log 10 of the viral load}
\usage{
vl_log(hiv1_rna)
}
\arguments{
\item{hiv1_rna}{Numeric. Viral Load. count/ml.}
}
\value{
Numeric. The log 10 of the hiv1_rna.
}
\description{
Log 10 of the viral load
}
\seealso{
Other utils: 
\code{\link{anemia}()},
\code{\link{bmi}()},
\code{\link{egfr}()},
\code{\link{fib4}()},
\code{\link{low_bmi}()}
}
\concept{utils}
