% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{anemia}
\alias{anemia}
\title{Anemia}
\usage{
anemia(hgb, sex)
}
\arguments{
\item{hgb}{Numeric. Hemoglobine. g/dL.}

\item{sex}{Character.}
}
\value{
Logical. TRUE if hgb < 13 g/dL. For female, TRUE if hgb < 12 g/dL.
}
\description{
Anemia
}
\seealso{
Other utils: 
\code{\link{bmi}()},
\code{\link{egfr}()},
\code{\link{fib4}()},
\code{\link{low_bmi}()},
\code{\link{trim_num}()},
\code{\link{vl_log}()}
}
\concept{utils}
