% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_num}
\alias{trim_num}
\title{Trim numerical vector}
\usage{
trim_num(x, lowerl = NULL, upperl = NULL)
}
\arguments{
\item{x}{Numeric, vector.}

\item{lowerl}{Numeric, scalar. Lower limit.}

\item{upperl}{Numeric, scalar, Upper limit.}
}
\value{
a numerical vector
}
\description{
Trim numerical vector
}
\keyword{internal}
