% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacsindex1.R
\name{vacsindex1}
\alias{vacsindex1}
\alias{score_vacsindex1}
\alias{points_vacsindex1}
\title{VACS Index 1}
\usage{
score_vacsindex1(age, cd4_count, hiv1_rna, hgb, fib4, egfr, hepc)

points_vacsindex1()
}
\arguments{
\item{age}{Numeric. Age in years.}

\item{cd4_count}{Numeric. CD4 cell count. cell/mm\eqn{^3}.}

\item{hiv1_rna}{Numeric. Viral load. copies/mL.}

\item{hgb}{Numeric. Hemoglobin. g/dL.}

\item{fib4}{Numeric. Fibrosis-4 Index for liver fibrosis.}

\item{egfr}{Numeric. Estimated glomerular filtration rate. (ml/min/1.73m\eqn{^2}).}

\item{hepc}{Logical. TRUE for co-infection with hepatitis C virus.}
}
\value{
\code{score_vacsindex1()} returns a numerical vector and \code{points_vacsindex1()} returns a list of the parameters used to compute the VACS Index 1.
}
\description{
Compute the VACS Index 1 given the predictors.
}
\examples{
data("vacs1")
vacs1$score_vacs1 <- score_vacsindex1(
    age = vacs1$age, cd4_count = vacs1$cd4, hiv1_rna = vacs1$hiv1_rna,
    hgb = vacs1$hgb, fib4 = vacs1$fib4, egfr = vacs1$egfr,
    hepc = vacs1$hepc)
}
\references{
Tate, J. P., Justice, A. C., Hughes, M. D., Bonnet, F., Reiss, P., Mocroft, A., ... & Sterne, J. A. (2013). An internationally generalizable risk index for mortality after one year of antiretroviral therapy. AIDS (London, England), 27(4), 563.

Bebu, I., Tate, J., Rimland, D., Mesner, O., Macalino, G. E., Ganesan, A., ... & Infectious Disease Clinical Research Program HIV Working Group. (2014). The VACS index predicts mortality in a young, healthy HIV population starting highly active antiretroviral therapy. Journal of acquired immune deficiency syndromes (1999), 65(2), 226.

Justice, A. C., Modur, S., Tate, J. P., Althoff, K. N., Jacobson, L. P., Gebo, K., ... & Gange, S. J. (2013). Predictive accuracy of the Veterans Aging Cohort Study (VACS) index for mortality with HIV infection: a north American cross cohort analysis. Journal of acquired immune deficiency syndromes (1999), 62(2), 149.
}
\concept{vacsindex1}
